+++ src/components/Accounts/Menu.js
@@ -150,6 +150,15 @@ const Menu = (props) => {
       >
         <FormattedMessage id="ui-users.accounts.history.button.transfer" />
       </Button>
+      <Button
+        id="fee-fine-report-export-button"
+        marginBottom0
+        disabled={_.isEmpty(feeFineActions)}
+        buttonStyle="primary"
+        onClick={props.onExportFeesFinesReport}
+      >
+        <FormattedMessage id="ui-users.export.button" />
+      </Button>
     </div>);
 
   return (
@@ -172,6 +181,7 @@ Menu.propTypes = {
   selectedAccounts: PropTypes.arrayOf(PropTypes.object).isRequired,
   feeFineActions: PropTypes.arrayOf(PropTypes.object).isRequired,
   onChangeActions: PropTypes.func,
+  onExportFeesFinesReport: PropTypes.func.isRequired,
   resources: PropTypes.object.isRequired,
 };